requirements.txt
asgiref==3.5.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.3
django-jalali-date==1.0.1
django-render-partial==0.4
django-widget-tweaks==1.4.12
idna==3.4
isodate==0.6.1
jdatetime==4.1.0
lxml==4.9.2
Pillow==9.3.0
platformdirs==2.6.0
pytz==2022.7
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.10.1
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.13
zeep==4.2.1
